+++ angular-workspace/projects/ni/nimble-angular/src/directives/tab-panel/nimble-tab-panel.directive.ts
@@ -1,7 +1,8 @@
 import { Directive } from '@angular/core';
-import type { TabPanel } from '@ni/nimble-components/dist/esm/tab-panel';
+import { type TabPanel, tabPanelTag } from '@ni/nimble-components/dist/esm/tab-panel';
 
 export type { TabPanel };
+export { tabPanelTag };
 
 /**
  * Directive to provide Angular integration for the tab panel.