+++ angular-workspace/projects/ni/nimble-angular/table-column/anchor/nimble-table-column-anchor.directive.ts
@@ -1,10 +1,11 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
-import type { TableColumnAnchor } from '@ni/nimble-components/dist/esm/table-column/anchor';
+import { type TableColumnAnchor, tableColumnAnchorTag } from '@ni/nimble-components/dist/esm/table-column/anchor';
 import { AnchorAppearance } from '@ni/nimble-components/dist/esm/anchor/types';
 import { BooleanValueOrAttribute, NumberValueOrAttribute, toBooleanProperty, toNullableNumberProperty } from '@ni/nimble-angular/internal-utilities';
 import { NimbleTableColumnBaseDirective } from '@ni/nimble-angular/table-column';
 
 export type { TableColumnAnchor };
+export { tableColumnAnchorTag };
 export { AnchorAppearance };
 
 /**
@@ -32,14 +33,6 @@ export class NimbleTableColumnAnchorDirective extends NimbleTableColumnBaseDirec
     // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('href-field-name') public set hrefFieldName(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'hrefFieldName', value);
-    }
-
-    public get placeholder(): string | undefined {
-        return this.elementRef.nativeElement.placeholder;
-    }
-
-    @Input() public set placeholder(value: string | undefined) {
-        this.renderer.setProperty(this.elementRef.nativeElement, 'placeholder', value);
     }
 
     public get appearance(): AnchorAppearance {