+++ packages/nimble-components/src/rich-text-viewer/tests/rich-text-viewer.stories.ts
@@ -9,7 +9,7 @@ import { richTextViewerTag } from '..';
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface RichTextViewerArgs {}
 
-const richTextViewerDescription = 'The rich text viewer component allows users to view text formatted with various styling options including bold, italics, numbered lists, and bulleted lists. The rich text to render is provided as a markdown string.';
+const richTextViewerDescription = 'The rich text viewer component allows users to view text formatted with various styling options including bold, italics, numbered lists, and bulleted lists. The rich text to render is provided as a markdown string.\n\n See the [rich text editor](?path=/docs/incubating-rich-text-editor--docs) component to enable users to modify the markdown contents.';
 
 const metadata: Meta<RichTextViewerArgs> = {
     title: 'Incubating/Rich Text Viewer',