+++ packages/nimble-components/src/table-column/date-text/cell-view/index.ts
@@ -1,0 +1,51 @@
+import { DesignSystem } from '@microsoft/fast-foundation';
+import { template } from '../../text-base/cell-view/template';
+import type {
+    TableColumnDateTextCellRecord,
+    TableColumnDateTextColumnConfig
+} from '..';
+import { styles } from '../../text-base/cell-view/styles';
+import { TableColumnTextCellViewBase } from '../../text-base/cell-view';
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'nimble-table-column-date-text-cell-view': TableColumnDateTextCellView;
+    }
+}
+
+/**
+ * A cell view for displaying date/time fields as text
+ */
+export class TableColumnDateTextCellView extends TableColumnTextCellViewBase<
+TableColumnDateTextCellRecord,
+TableColumnDateTextColumnConfig
+> {
+    private static readonly formatter = new Intl.DateTimeFormat(undefined, {
+        dateStyle: 'medium',
+        timeStyle: 'medium'
+    });
+
+    private cellRecordChanged(): void {
+        if (typeof this.cellRecord.value === 'number') {
+            try {
+                this.text = TableColumnDateTextCellView.formatter.format(
+                    this.cellRecord.value
+                );
+            } catch (e) {
+                this.text = '';
+            }
+        } else {
+            this.text = '';
+        }
+    }
+}
+
+const dateTextCellView = TableColumnDateTextCellView.compose({
+    baseName: 'table-column-date-text-cell-view',
+    template,
+    styles
+});
+DesignSystem.getOrCreate().withPrefix('nimble').register(dateTextCellView());
+export const tableColumnDateTextCellViewTag = DesignSystem.tagFor(
+    TableColumnDateTextCellView
+);